Test: Refactor Test Directory
